refactor tax provider inheritance for aggregator scope
